requirements.txt
mido==1.3.0
packaging==23.2
pygame==2.5.1
python-rtmidi==1.5.5
